Fix milestone block numbers in genesis.json
The block numbers for milestones in config.json should be all "0". Or if absolutely required to be non-zero, they have to be ordered in the real temporal sequence that they happened in the Ethereum mainnet.
For example, **byzantiumBlock** can not specify a block number lower than **homesteadBlock**.

Some Ethereum clients may not check compliance, but others like Pantheon check the consistency of the block numbers for milestones, and reject to use this genesis.json files.